requirements.txt
blinker==1.4
click==7.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.1
PyJWT==1.7.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
